bench/requirements.txt
litellm>=1.60
pyyaml>=6.0
